Move test to test_ir. Add doc-string
